requirements.txt
git+https://github.com/sndrtj/pyrefflat.git@master
flask
flask_restful
requests
flask_bootstrap
flask-wtf
flask-table
flask-restful-swagger
Flask-SQLAlchemy
Flask-Login
Flask-Restful
Flask-Table
Jinja2
Flask-WTF
suds
pybedtools
raven[flask]
flask_jsglue